Changed a couple of features such as making the grade stored as 16 bit uint instead of 8 bits because I remebered that I'm dumb and need another byte for the full range of atleast 100, not to mention any extra/bonus points. I also added a basic addGroup command option in command line

+++ include/groups/Class.hpp
@@ -9,21 +9,21 @@
 
 class Class : public Group {
     private:
-        uint8_t year;
+        uint16_t year;
         Semester semester;
         Topic topic;
-        int grade;
+        uint8_t grade;
     public:
         Class(
             int idNum, 
             const std::string& name, 
-            uint8_t year, 
+            uint16_t year, 
             const Semester& semester, 
             const Topic& topic,
             uint8_t grade
         );
-        uint8_t getYear() const;
-        void setYear(uint8_t year);
+        uint16_t getYear() const;
+        void setYear(uint16_t year);
         Semester getSemester() const;
         void setSemester(const Semester& semester);
         Topic getTopic() const;
